bat_pytest/requirements.txt
allure-pytest==2.12.0
pytest==7.2.1
boto3
pytest-logging==2015.11.4